EOS-16854-3rd_party_sw_monitoring_new (ThreadException class moved from thread_exception.py to module_thread.py)
