+++ low-level/framework/base/module_thread.py
@@ -286,4 +286,14 @@ class SensorThread(ScheduledModuleThread):
         else:
             self._scheduler.enter(1, self._priority, self.run, ())
 
+class ThreadException(Exception):
+    """Generic Exception to handle Threads errors."""
 
+    def __init__(self, module, message):
+        """Handle error msg from thread modules."""
+        self._module = module
+        self._desc = message
+
+    def __str__(self):
+        """Returns formated error msg."""
+        return "%s: error: %s" %(self._module, self._desc)
